+++ allauth/account/adapter.py
@@ -25,8 +25,8 @@ from django.template import TemplateDoesNotExist
 from django.template.loader import render_to_string
 from django.urls import reverse
 from django.utils import timezone
-
-from allauth.compat import force_str, ugettext_lazy as _
+from django.utils.encoding import force_str
+from django.utils.translation import gettext_lazy as _
 
 from ..utils import (
     build_absolute_uri,