+++ allauth/socialaccount/adapter.py
@@ -2,8 +2,7 @@ from __future__ import absolute_import
 
 from django.core.exceptions import ValidationError
 from django.urls import reverse
-
-from allauth.compat import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from ..account import app_settings as account_settings
 from ..account.adapter import get_adapter as get_account_adapter
@@ -190,6 +189,19 @@ class DefaultSocialAccountAdapter(object):
 
     def serialize_instance(self, instance):
         return serialize_instance(instance)
+
+    def get_app(self, request, provider):
+        # NOTE: Avoid loading models at top due to registry boot...
+        from allauth.socialaccount.models import SocialApp
+
+        config = app_settings.PROVIDERS.get(provider, {}).get('APP')
+        if config:
+            app = SocialApp(provider=provider)
+            for field in ['client_id', 'secret', 'key']:
+                setattr(app, field, config.get(field))
+        else:
+            app = SocialApp.objects.get_current(provider, request)
+        return app
 
 
 def get_adapter(request=None):