+++ allauth/socialaccount/providers/amazon_cognito/utils.py
@@ -1,0 +1,7 @@
+def convert_to_python_bool_if_value_is_json_string_bool(s):
+    if s == "true":
+        return True
+    elif s == "false":
+        return False
+
+    return s