+++ allauth/socialaccount/providers/steam/urls.py
@@ -1,9 +1,9 @@
-from django.conf.urls import url
+from django.urls import path
 
 from . import views
 
 
 urlpatterns = [
-    url("^steam/login/$", views.steam_login, name="steam_login"),
-    url("^steam/callback/$", views.steam_callback, name="steam_callback"),
+    path("steam/login/", views.steam_login, name="steam_login"),
+    path("steam/callback/", views.steam_callback, name="steam_callback"),
 ]