+++ allauth/socialaccount/urls.py
@@ -1,13 +1,13 @@
-from django.conf.urls import url
+from django.urls import path
 
 from . import views
 
 
 urlpatterns = [
-    url(r'^login/cancelled/$', views.login_cancelled,
-        name='socialaccount_login_cancelled'),
-    url(r'^login/error/$', views.login_error,
-        name='socialaccount_login_error'),
-    url(r'^signup/$', views.signup, name='socialaccount_signup'),
-    url(r'^connections/$', views.connections, name='socialaccount_connections')
+    path('login/cancelled/', views.login_cancelled,
+         name='socialaccount_login_cancelled'),
+    path('login/error/', views.login_error,
+         name='socialaccount_login_error'),
+    path('signup/', views.signup, name='socialaccount_signup'),
+    path('connections/', views.connections, name='socialaccount_connections')
 ]