+++ allauth/urls.py
@@ -1,16 +1,16 @@
 from importlib import import_module
 
-from django.conf.urls import include, url
+from django.urls import include, path
 
 from allauth.socialaccount import providers
 
 from . import app_settings
 
 
-urlpatterns = [url(r'^', include('allauth.account.urls'))]
+urlpatterns = [path('', include('allauth.account.urls'))]
 
 if app_settings.SOCIALACCOUNT_ENABLED:
-    urlpatterns += [url(r'^social/', include('allauth.socialaccount.urls'))]
+    urlpatterns += [path('social/', include('allauth.socialaccount.urls'))]
 
 # Provider urlpatterns, as separate attribute (for reusability).
 provider_urlpatterns = []