+++ example/example/demo/apps.py
@@ -1,7 +1,6 @@
 from django.apps import AppConfig
 from django.db.models.signals import post_migrate
-
-from allauth.compat import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 
 def setup_dummy_social_apps(sender, **kwargs):
@@ -23,10 +22,10 @@ def setup_dummy_social_apps(sender, **kwargs):
                 SocialApp.objects.get(provider=provider.id,
                                       sites=site)
             except SocialApp.DoesNotExist:
-                print ("Installing dummy application credentials for %s."
-                       " Authentication via this provider will not work"
-                       " until you configure proper credentials via the"
-                       " Django admin (`SocialApp` models)" % provider.id)
+                print("Installing dummy application credentials for %s."
+                      " Authentication via this provider will not work"
+                      " until you configure proper credentials via the"
+                      " Django admin (`SocialApp` models)" % provider.id)
                 app = SocialApp.objects.create(
                     provider=provider.id,
                     secret='secret',